https://github.com/WWBN/AVideo/issues/7554

+++ objects/Object.php
@@ -636,7 +636,9 @@ abstract class ObjectYPT implements ObjectInterface
             AVideoPlugin::loadPlugin('Cache');
         }
 
-        Cache::deleteCache($name);
+        if (class_exists('Cache')) {
+            Cache::deleteCache($name);
+        }
 
         global $__getAVideoCache;
         unset($__getAVideoCache);
