+++ objects/plugin.php
@@ -176,7 +176,7 @@ class Plugin extends ObjectYPT
                 $getPluginByUUID[$uuid] = $data;
             } else {
                 $name = AVideoPlugin::getPluginsNameOnByDefaultFromUUID($uuid);
-                if ($name !== false && empty($pluginJustInstalled[$uuid])) {
+                if (!sqlDAL::wasSTMTError() && $name !== false && empty($pluginJustInstalled[$uuid])) {
                     $pluginJustInstalled[$uuid] = 1;
                     _error_log("plugin::getPluginByUUID {$name} {$uuid} this plugin is On By Default we will install it ($sql)");
                     self::deleteByUUID($uuid);
